+++ application/extractor/extractor_test.go
@@ -25,66 +25,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// node.go
-func TestNodeInfoResponse(t *testing.T) {
-	testPK := "test_public_key"
-	testRole := core.StaticRoleVirtual
-
-	testValue := struct {
-		PublicKey string
-		Role      core.StaticRole
-	}{
-		PublicKey: testPK,
-		Role:      testRole,
-	}
-
-	data, err := core.Serialize([]interface{}{testValue, nil})
-	require.NoError(t, err)
-
-	pk, role, err := NodeInfoResponse(data)
-
-	require.NoError(t, err)
-	require.Equal(t, testPK, pk)
-	require.Equal(t, testRole.String(), role)
-}
-
-func TestNodeInfoResponse_ErrorResponse(t *testing.T) {
-	testPK := "test_public_key"
-	testRole := core.StaticRoleVirtual
-
-	testValue := struct {
-		PublicKey string
-		Role      core.StaticRole
-	}{
-		PublicKey: testPK,
-		Role:      testRole,
-	}
-	contractErr := &foundation.Error{S: "Custom test error"}
-
-	data, err := core.Serialize([]interface{}{testValue, contractErr})
-	require.NoError(t, err)
-
-	pk, role, err := NodeInfoResponse(data)
-
-	require.Contains(t, err.Error(), "Has error in response")
-	require.Contains(t, err.Error(), "Custom test error")
-	require.Equal(t, "", pk)
-	require.Equal(t, "", role)
-}
-
-func TestNodeInfoResponse_UnmarshalError(t *testing.T) {
-	testValue := "some_no_valid_data"
-
-	data, err := core.Serialize(testValue)
-	require.NoError(t, err)
-
-	pk, role, err := NodeInfoResponse(data)
-
-	require.Contains(t, err.Error(), "Can't unmarshal response")
-	require.Equal(t, "", pk)
-	require.Equal(t, "", role)
-}
-
 // rootdomain.go
 func TestInfoResponse(t *testing.T) {
 	testValue, _ := json.Marshal(map[string]interface{}{
